feat: useIsModalOpen hook 분리 #50
